F81 Fix UV index float (not int)

+++ db/migrate/20171204012402_create_uv_indices.rb
@@ -2,7 +2,7 @@ class CreateUvIndices < ActiveRecord::Migration[5.1]
   def change
     create_table :uv_indices do |t|
       t.datetime :date_time, null: false
-      t.integer :uv_index
+      t.float :uv_index
       t.string :scale
       t.integer :spot_id, null: false
